+++ app/models/business_subscription_checkout.rb
@@ -19,13 +19,13 @@ class BusinessSubscriptionCheckout
     user.set_payment_processor(:stripe)
     user.payment_processor.checkout(
       mode: "subscription",
-      line_items: plan.price_id,
+      line_items: plan.stripe_price_id,
       success_url: analytics_event_url(event)
     )
   end
 
   def plan
-    Businesses::Subscription.with_identifier(@plan)
+    Businesses::Plan.with_identifier(@plan)
   end
 
   def event