+++ app/models/feature.rb
@@ -4,7 +4,7 @@ class Feature
     when :new_developer_fields_banner
       true
     when :message_read_indicator
-      !Rails.env.production? || user&.admin?
+      true
     when :pricing_v2
       true
     when :reply_via_email