+++ config/application.rb
@@ -1,19 +1,11 @@
 require_relative "boot"
+require_relative "subscriptions"
 
 require "rails/all"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
-
-def stripe_price_id(subscription)
-  credentials = Rails.application.credentials
-  if Rails.env.development?
-    credentials.dig(:stripe, :price_ids, subscription) || "#{subscription}_dummy_price_id"
-  else
-    credentials.stripe[:price_ids][subscription]
-  end
-end
 
 module Railsdevs
   class Application < Rails::Application
@@ -36,8 +28,8 @@ module Railsdevs
     config.analytics = config_for(:analytics)
     config.emails = config_for(:emails)
     config.fathom = config_for(:fathom)
+    config.plans = config_for(:plans)
     config.sitemaps_host = "https://#{Rails.application.credentials.dig(:aws, :sitemaps_bucket)}.s3.#{Rails.application.credentials.dig(:aws, :region)}.amazonaws.com/"
-    config.subscriptions = config_for(:subscriptions)
     config.upload_sitemap = false
 
     # Run background jobs via sidekiq.