+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## 2022
 
 ### July
-* July 2 - Add developer and business avatar images from Unsplash - #530 @adrienpoly
+* July 10 - Seed developer and business avatar images from Unsplash - #530 @adrienpoly
+* July 10 - Add privacy policy and terms & conditions
+* July 9 - In-app purchase support on iOS via RevenueCat
+* July 7 - Impersonate users as an admin
 * July 1 - Fixes a bug where creating an invisible developer would raise an error - #534 @joshio1
 
 ### June